A little refactoring of the parser
This introduces property functions for hasText and currentCharacter. hasText replaces and reverses the logic for eof(). The NFA ‘join’ function has been moved to NFAGraph, which makes a little more sense to me.

This also allows us to optionally ignore whitespace in regular expressions. Previously this was forced. Now this behavior can be changed with the ignoreWhitespace flag.